refactor: split large rust modules and extract tests
