--- src/sql/session.rs
@@ -1,1389 +1,0 @@
-use crate::error::{MuroError, Result};
-use crate::schema::catalog::SystemCatalog;
-use crate::sql::ast::Statement;
-use crate::sql::executor::{execute_statement, ExecResult, Row};
-use crate::sql::parser::parse_sql;
-use crate::storage::pager::Pager;
-use crate::tx::page_store::TxPageStore;
-use crate::tx::transaction::Transaction;
-use crate::types::Value;
-use crate::wal::record::TxId;
-use crate::wal::writer::WalWriter;
-
-const CHECKPOINT_MAX_ATTEMPTS: usize = 2;
-const DEFAULT_CHECKPOINT_TX_THRESHOLD: u64 = 1;
-const DEFAULT_CHECKPOINT_WAL_BYTES_THRESHOLD: u64 = 0;
-const DEFAULT_CHECKPOINT_INTERVAL_MS: u64 = 0;
-
-#[derive(Debug, Clone, Copy)]
-struct CheckpointPolicy {
-    tx_threshold: u64,
-    wal_bytes_threshold: u64,
-    interval_ms: u64,
-}
-
-impl Default for CheckpointPolicy {
-    fn default() -> Self {
-        Self {
-            tx_threshold: DEFAULT_CHECKPOINT_TX_THRESHOLD,
-            wal_bytes_threshold: DEFAULT_CHECKPOINT_WAL_BYTES_THRESHOLD,
-            interval_ms: DEFAULT_CHECKPOINT_INTERVAL_MS,
-        }
-    }
-}
-
-impl CheckpointPolicy {
-    fn from_env() -> Self {
-        Self {
-            tx_threshold: parse_checkpoint_env_u64(
-                "MURODB_CHECKPOINT_TX_THRESHOLD",
-                DEFAULT_CHECKPOINT_TX_THRESHOLD,
-                0,
-            ),
-            wal_bytes_threshold: parse_checkpoint_env_u64(
-                "MURODB_CHECKPOINT_WAL_BYTES_THRESHOLD",
-                DEFAULT_CHECKPOINT_WAL_BYTES_THRESHOLD,
-                0,
-            ),
-            interval_ms: parse_checkpoint_env_u64(
-                "MURODB_CHECKPOINT_INTERVAL_MS",
-                DEFAULT_CHECKPOINT_INTERVAL_MS,
-                0,
-            ),
-        }
-    }
-}
-
-fn parse_checkpoint_env_u64(name: &str, default: u64, min: u64) -> u64 {
-    let Ok(raw) = std::env::var(name) else {
-        return default;
-    };
-    match raw.parse::<u64>() {
-        Ok(v) if v >= min => v,
-        Ok(_) => {
-            eprintln!(
-                "WARNING: {} must be >= {}, using default {}",
-                name, min, default
-            );
-            default
-        }
-        Err(_) => {
-            eprintln!(
-                "WARNING: {} must be an integer, using default {}",
-                name, default
-            );
-            default
-        }
-    }
-}
-
-/// Database operation statistics for observability.
-#[derive(Debug, Clone, Default)]
-pub struct DatabaseStats {
-    // Checkpoint stats
-    pub total_checkpoints: u64,
-    pub failed_checkpoints: u64,
-    pub last_failure_error: Option<String>,
-    pub last_failure_timestamp_ms: Option<u64>,
-    // CommitInDoubt stats
-    pub commit_in_doubt_count: u64,
-    pub last_commit_in_doubt_error: Option<String>,
-    pub last_commit_in_doubt_timestamp_ms: Option<u64>,
-    // Freelist sanitize stats
-    pub freelist_sanitize_count: u64,
-    pub freelist_out_of_range_total: u64,
-    pub freelist_duplicates_total: u64,
-    pub deferred_checkpoints: u64,
-}
-
-/// Backward-compatible alias.
-pub type CheckpointStats = DatabaseStats;
-
-/// A session that manages explicit transaction state.
-///
-/// - `BEGIN` starts a transaction (dirty-page buffering).
-/// - `COMMIT` flushes dirty pages via WAL, then to disk.
-/// - `ROLLBACK` discards dirty pages without WAL append.
-/// - Without `BEGIN`, each statement executes in auto-commit mode
-///   (wrapped in an implicit transaction with WAL).
-pub struct Session {
-    pager: Pager,
-    catalog: SystemCatalog,
-    wal: WalWriter,
-    active_tx: Option<Transaction>,
-    next_txid: TxId,
-    stats: DatabaseStats,
-    poisoned: Option<String>,
-    checkpoint_policy: CheckpointPolicy,
-    pending_checkpoint_ops: u64,
-    last_checkpoint_at: std::time::Instant,
-    #[cfg(test)]
-    inject_checkpoint_failures_remaining: usize,
-}
-
-impl Session {
-    pub fn new(pager: Pager, catalog: SystemCatalog, wal: WalWriter) -> Self {
-        let next_txid = pager.next_txid();
-        let mut stats = DatabaseStats::default();
-
-        // Absorb freelist sanitize report from pager open
-        if let Some(report) = pager.freelist_sanitize_report() {
-            stats.freelist_sanitize_count += 1;
-            stats.freelist_out_of_range_total += report.out_of_range.len() as u64;
-            stats.freelist_duplicates_total += report.duplicates.len() as u64;
-            eprintln!(
-                "WARNING: freelist_sanitized out_of_range={} duplicates={} total_removed={}",
-                report.out_of_range.len(),
-                report.duplicates.len(),
-                report.total_removed()
-            );
-        }
-
-        Session {
-            pager,
-            catalog,
-            wal,
-            active_tx: None,
-            next_txid,
-            stats,
-            poisoned: None,
-            checkpoint_policy: CheckpointPolicy::from_env(),
-            pending_checkpoint_ops: 0,
-            last_checkpoint_at: std::time::Instant::now(),
-            #[cfg(test)]
-            inject_checkpoint_failures_remaining: 0,
-        }
-    }
-
-    fn check_poisoned(&self) -> Result<()> {
-        if let Some(ref msg) = self.poisoned {
-            return Err(MuroError::SessionPoisoned(msg.clone()));
-        }
-        Ok(())
-    }
-
-    /// Execute a SQL string, handling BEGIN/COMMIT/ROLLBACK at the session level.
-    pub fn execute(&mut self, sql: &str) -> Result<ExecResult> {
-        let stmt = parse_sql(sql).map_err(MuroError::Parse)?;
-
-        // Stats queries are always allowed, even on poisoned sessions,
-        // so operators can inspect counters after CommitInDoubt.
-        match &stmt {
-            Statement::ShowCheckpointStats => return self.handle_show_checkpoint_stats(),
-            Statement::ShowDatabaseStats => return self.handle_show_database_stats(),
-            _ => {}
-        }
-
-        self.check_poisoned()?;
-        self.refresh_from_disk_if_needed()?;
-
-        match &stmt {
-            Statement::Begin => self.handle_begin(),
-            Statement::Commit => self.handle_commit(),
-            Statement::Rollback => self.handle_rollback(),
-            _ => {
-                if self.active_tx.is_some() {
-                    self.execute_in_tx(&stmt)
-                } else {
-                    // Auto-commit: wrap in an implicit transaction with WAL
-                    self.execute_auto_commit(&stmt)
-                }
-            }
-        }
-    }
-
-    /// Execute a read-only SQL query and return rows.
-    ///
-    /// This path avoids auto-commit WAL writes for non-transactional reads.
-    pub fn execute_read_only_query(&mut self, sql: &str) -> Result<Vec<Row>> {
-        let stmt = parse_sql(sql).map_err(MuroError::Parse)?;
-
-        // Stats queries are always allowed, even on poisoned sessions.
-        match &stmt {
-            Statement::ShowCheckpointStats => {
-                return Self::rows_from_exec_result(self.handle_show_checkpoint_stats())
-            }
-            Statement::ShowDatabaseStats => {
-                return Self::rows_from_exec_result(self.handle_show_database_stats())
-            }
-            _ => {}
-        }
-
-        self.check_poisoned()?;
-        self.refresh_from_disk_if_needed()?;
-
-        if !Self::is_read_only_statement(&stmt) {
-            return Err(MuroError::Execution(
-                "Database::query accepts read-only SQL only; use execute() for writes".into(),
-            ));
-        }
-
-        if self.active_tx.is_some() {
-            Self::rows_from_exec_result(self.execute_in_tx(&stmt))
-        } else {
-            // Read directly from pager/catalog without opening an implicit WAL transaction.
-            Self::rows_from_exec_result(execute_statement(
-                &stmt,
-                &mut self.pager,
-                &mut self.catalog,
-            ))
-        }
-    }
-
-    fn rows_from_exec_result(result: Result<ExecResult>) -> Result<Vec<Row>> {
-        match result? {
-            ExecResult::Rows(rows) => Ok(rows),
-            ExecResult::RowsAffected(_) | ExecResult::Ok => Err(MuroError::Execution(
-                "Read-only query must return rows".into(),
-            )),
-        }
-    }
-
-    fn is_read_only_statement(stmt: &Statement) -> bool {
-        match stmt {
-            Statement::Select(_)
-            | Statement::SetQuery(_)
-            | Statement::ShowTables
-            | Statement::ShowCreateTable(_)
-            | Statement::Describe(_)
-            | Statement::ShowCheckpointStats
-            | Statement::ShowDatabaseStats => true,
-            Statement::Explain(inner) => Self::is_read_only_statement(inner),
-            Statement::CreateTable(_)
-            | Statement::CreateIndex(_)
-            | Statement::CreateFulltextIndex(_)
-            | Statement::DropTable(_)
-            | Statement::DropIndex(_)
-            | Statement::AlterTable(_)
-            | Statement::RenameTable(_)
-            | Statement::Insert(_)
-            | Statement::Update(_)
-            | Statement::Delete(_)
-            | Statement::AnalyzeTable(_)
-            | Statement::Begin
-            | Statement::Commit
-            | Statement::Rollback => false,
-        }
-    }
-
-    fn refresh_from_disk_if_needed(&mut self) -> Result<()> {
-        if self.active_tx.is_some() {
-            return Ok(());
-        }
-        if self.pager.refresh_from_disk_if_changed()? {
-            self.catalog = SystemCatalog::open(self.pager.catalog_root());
-            self.next_txid = self.next_txid.max(self.pager.next_txid());
-        }
-        Ok(())
-    }
-
-    fn handle_begin(&mut self) -> Result<ExecResult> {
-        if self.active_tx.is_some() {
-            return Err(MuroError::Transaction("Transaction already active".into()));
-        }
-        let txid = self.next_txid;
-        self.next_txid += 1;
-        let snapshot_lsn = self.wal.current_lsn();
-        self.active_tx = Some(Transaction::begin(txid, snapshot_lsn));
-        Ok(ExecResult::Ok)
-    }
-
-    fn handle_commit(&mut self) -> Result<ExecResult> {
-        let mut tx = self
-            .active_tx
-            .take()
-            .ok_or_else(|| MuroError::Transaction("No active transaction".into()))?;
-        let catalog_root = self.catalog.root_page_id();
-        self.pager.set_next_txid(self.next_txid);
-        match tx.commit(&mut self.pager, &mut self.wal, catalog_root) {
-            Err(e @ MuroError::CommitInDoubt(_)) => {
-                self.record_commit_in_doubt(&e);
-                self.poisoned = Some(e.to_string());
-                return Err(e);
-            }
-            Err(e) => return Err(e),
-            Ok(_) => {}
-        }
-        self.post_commit_checkpoint();
-        Ok(ExecResult::Ok)
-    }
-
-    fn handle_rollback(&mut self) -> Result<ExecResult> {
-        let mut tx = self
-            .active_tx
-            .take()
-            .ok_or_else(|| MuroError::Transaction("No active transaction".into()))?;
-        tx.rollback_no_wal();
-        self.post_rollback_checkpoint();
-        // Reload catalog from disk since in-memory catalog may have been modified
-        let catalog_root = self.pager.catalog_root();
-        self.catalog = SystemCatalog::open(catalog_root);
-        Ok(ExecResult::Ok)
-    }
-
-    /// Execute a statement in auto-commit mode: wrap in an implicit transaction.
-    fn execute_auto_commit(&mut self, stmt: &Statement) -> Result<ExecResult> {
-        let txid = self.next_txid;
-        self.next_txid += 1;
-        let snapshot_lsn = self.wal.current_lsn();
-        let tx = Transaction::begin(txid, snapshot_lsn);
-
-        // Save catalog state for rollback on error
-        let catalog_root_before = self.catalog.root_page_id();
-
-        let mut store = TxPageStore::new(tx, &mut self.pager);
-        let result = execute_statement(stmt, &mut store, &mut self.catalog);
-        let mut tx = store.into_tx();
-
-        match result {
-            Ok(exec_result) => {
-                // Commit via WAL (catalog_root included in WAL MetaUpdate)
-                let catalog_root = self.catalog.root_page_id();
-                self.pager.set_next_txid(self.next_txid);
-                match tx.commit(&mut self.pager, &mut self.wal, catalog_root) {
-                    Err(e @ MuroError::CommitInDoubt(_)) => {
-                        self.record_commit_in_doubt(&e);
-                        self.poisoned = Some(e.to_string());
-                        return Err(e);
-                    }
-                    Err(e) => return Err(e),
-                    Ok(_) => {}
-                }
-                self.post_commit_checkpoint();
-                Ok(exec_result)
-            }
-            Err(e) => {
-                // Rollback: discard dirty pages, restore catalog
-                tx.rollback_no_wal();
-                self.catalog = SystemCatalog::open(catalog_root_before);
-                Err(e)
-            }
-        }
-    }
-
-    /// Execute a statement within an active transaction.
-    fn execute_in_tx(&mut self, stmt: &Statement) -> Result<ExecResult> {
-        // Save catalog state so we can restore on error
-        let catalog_root_before = self.catalog.root_page_id();
-
-        // Take the transaction out temporarily
-        let tx = self.active_tx.take().unwrap();
-        let mut store = TxPageStore::new(tx, &mut self.pager);
-
-        let result = execute_statement(stmt, &mut store, &mut self.catalog);
-
-        // Put the transaction back
-        self.active_tx = Some(store.into_tx());
-
-        if result.is_err() {
-            // Restore catalog to pre-statement state on error
-            self.catalog = SystemCatalog::open(catalog_root_before);
-        }
-
-        result
-    }
-
-    /// Get a reference to the pager (for flush/metadata operations).
-    pub fn pager(&self) -> &Pager {
-        &self.pager
-    }
-
-    /// Get a mutable reference to the pager.
-    pub fn pager_mut(&mut self) -> &mut Pager {
-        &mut self.pager
-    }
-
-    /// Get a mutable reference to the WAL writer.
-    #[cfg(any(test, feature = "test-utils"))]
-    pub fn wal_mut(&mut self) -> &mut WalWriter {
-        &mut self.wal
-    }
-
-    /// Get a reference to the catalog.
-    pub fn catalog(&self) -> &SystemCatalog {
-        &self.catalog
-    }
-
-    fn post_commit_checkpoint(&mut self) {
-        self.pending_checkpoint_ops = self.pending_checkpoint_ops.saturating_add(1);
-        if !self.should_checkpoint_now() {
-            self.stats.deferred_checkpoints += 1;
-            return;
-        }
-        self.stats.total_checkpoints += 1;
-        // Best-effort: commit already reached durable state in data file.
-        // If WAL truncate fails, keep serving and rely on startup recovery path.
-        if let Err((attempts, e)) = self.try_checkpoint_truncate_with_retry() {
-            self.stats.failed_checkpoints += 1;
-            self.stats.last_failure_error = Some(format!("{}", e));
-            self.stats.last_failure_timestamp_ms = Some(
-                std::time::SystemTime::now()
-                    .duration_since(std::time::UNIX_EPOCH)
-                    .unwrap_or_default()
-                    .as_millis() as u64,
-            );
-            self.emit_checkpoint_warning("post-commit", attempts, &e);
-            return;
-        }
-        self.pending_checkpoint_ops = 0;
-        self.last_checkpoint_at = std::time::Instant::now();
-    }
-
-    fn post_rollback_checkpoint(&mut self) {
-        self.pending_checkpoint_ops = self.pending_checkpoint_ops.saturating_add(1);
-        if !self.should_checkpoint_now() {
-            self.stats.deferred_checkpoints += 1;
-            return;
-        }
-        self.stats.total_checkpoints += 1;
-        // Best-effort: rollback leaves no committed changes to preserve in WAL.
-        if let Err((attempts, e)) = self.try_checkpoint_truncate_with_retry() {
-            self.stats.failed_checkpoints += 1;
-            self.stats.last_failure_error = Some(format!("{}", e));
-            self.stats.last_failure_timestamp_ms = Some(
-                std::time::SystemTime::now()
-                    .duration_since(std::time::UNIX_EPOCH)
-                    .unwrap_or_default()
-                    .as_millis() as u64,
-            );
-            self.emit_checkpoint_warning("post-rollback", attempts, &e);
-            return;
-        }
-        self.pending_checkpoint_ops = 0;
-        self.last_checkpoint_at = std::time::Instant::now();
-    }
-
-    fn should_checkpoint_now(&self) -> bool {
-        if self.pending_checkpoint_ops == 0 {
-            return false;
-        }
-        if self.checkpoint_policy.tx_threshold > 0 {
-            if self.checkpoint_policy.tx_threshold <= 1 {
-                return true;
-            }
-            if self.pending_checkpoint_ops >= self.checkpoint_policy.tx_threshold {
-                return true;
-            }
-        }
-        if self.checkpoint_policy.wal_bytes_threshold > 0 {
-            if let Ok(size) = self.wal.file_size_bytes() {
-                if size >= self.checkpoint_policy.wal_bytes_threshold {
-                    return true;
-                }
-            }
-        }
-        if self.checkpoint_policy.interval_ms > 0
-            && self.last_checkpoint_at.elapsed().as_millis() as u64
-                >= self.checkpoint_policy.interval_ms
-        {
-            return true;
-        }
-        false
-    }
-
-    fn handle_show_checkpoint_stats(&self) -> Result<ExecResult> {
-        let stats = &self.stats;
-        let rows = vec![
-            Row {
-                values: vec![
-                    (
-                        "stat".to_string(),
-                        Value::Varchar("total_checkpoints".to_string()),
-                    ),
-                    (
-                        "value".to_string(),
-                        Value::Varchar(stats.total_checkpoints.to_string()),
-                    ),
-                ],
-            },
-            Row {
-                values: vec![
-                    (
-                        "stat".to_string(),
-                        Value::Varchar("failed_checkpoints".to_string()),
-                    ),
-                    (
-                        "value".to_string(),
-                        Value::Varchar(stats.failed_checkpoints.to_string()),
-                    ),
-                ],
-            },
-            Row {
-                values: vec![
-                    (
-                        "stat".to_string(),
-                        Value::Varchar("last_failure_error".to_string()),
-                    ),
-                    (
-                        "value".to_string(),
-                        Value::Varchar(stats.last_failure_error.clone().unwrap_or_default()),
-                    ),
-                ],
-            },
-            Row {
-                values: vec![
-                    (
-                        "stat".to_string(),
-                        Value::Varchar("last_failure_timestamp_ms".to_string()),
-                    ),
-                    (
-                        "value".to_string(),
-                        Value::Varchar(
-                            stats
-                                .last_failure_timestamp_ms
-                                .map(|v| v.to_string())
-                                .unwrap_or_default(),
-                        ),
-                    ),
-                ],
-            },
-        ];
-        Ok(ExecResult::Rows(rows))
-    }
-
-    pub fn checkpoint_stats(&self) -> &CheckpointStats {
-        &self.stats
-    }
-
-    pub fn database_stats(&self) -> &DatabaseStats {
-        &self.stats
-    }
-
-    fn record_commit_in_doubt(&mut self, error: &MuroError) {
-        self.stats.commit_in_doubt_count += 1;
-        self.stats.last_commit_in_doubt_error = Some(error.to_string());
-        self.stats.last_commit_in_doubt_timestamp_ms = Some(
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .unwrap_or_default()
-                .as_millis() as u64,
-        );
-        eprintln!("WARNING: commit_in_doubt error=\"{}\"", error);
-    }
-
-    fn handle_show_database_stats(&self) -> Result<ExecResult> {
-        let stats = &self.stats;
-        let cache_hits = self.pager.cache_hits();
-        let cache_misses = self.pager.cache_misses();
-        let cache_total = cache_hits.saturating_add(cache_misses);
-        let cache_hit_rate_pct = if cache_total == 0 {
-            0.0
-        } else {
-            (cache_hits as f64 * 100.0) / (cache_total as f64)
-        };
-        fn stat_row(name: &str, value: String) -> Row {
-            Row {
-                values: vec![
-                    ("stat".to_string(), Value::Varchar(name.to_string())),
-                    ("value".to_string(), Value::Varchar(value)),
-                ],
-            }
-        }
-        let rows = vec![
-            stat_row("total_checkpoints", stats.total_checkpoints.to_string()),
-            stat_row("failed_checkpoints", stats.failed_checkpoints.to_string()),
-            stat_row(
-                "last_failure_error",
-                stats.last_failure_error.clone().unwrap_or_default(),
-            ),
-            stat_row(
-                "last_failure_timestamp_ms",
-                stats
-                    .last_failure_timestamp_ms
-                    .map(|v| v.to_string())
-                    .unwrap_or_default(),
-            ),
-            stat_row(
-                "commit_in_doubt_count",
-                stats.commit_in_doubt_count.to_string(),
-            ),
-            stat_row(
-                "last_commit_in_doubt_error",
-                stats.last_commit_in_doubt_error.clone().unwrap_or_default(),
-            ),
-            stat_row(
-                "last_commit_in_doubt_timestamp_ms",
-                stats
-                    .last_commit_in_doubt_timestamp_ms
-                    .map(|v| v.to_string())
-                    .unwrap_or_default(),
-            ),
-            stat_row(
-                "freelist_sanitize_count",
-                stats.freelist_sanitize_count.to_string(),
-            ),
-            stat_row(
-                "freelist_out_of_range_total",
-                stats.freelist_out_of_range_total.to_string(),
-            ),
-            stat_row(
-                "freelist_duplicates_total",
-                stats.freelist_duplicates_total.to_string(),
-            ),
-            stat_row(
-                "deferred_checkpoints",
-                stats.deferred_checkpoints.to_string(),
-            ),
-            stat_row(
-                "checkpoint_pending_ops",
-                self.pending_checkpoint_ops.to_string(),
-            ),
-            stat_row(
-                "checkpoint_policy_tx_threshold",
-                self.checkpoint_policy.tx_threshold.to_string(),
-            ),
-            stat_row(
-                "checkpoint_policy_wal_bytes_threshold",
-                self.checkpoint_policy.wal_bytes_threshold.to_string(),
-            ),
-            stat_row(
-                "checkpoint_policy_interval_ms",
-                self.checkpoint_policy.interval_ms.to_string(),
-            ),
-            stat_row("pager_cache_hits", cache_hits.to_string()),
-            stat_row("pager_cache_misses", cache_misses.to_string()),
-            stat_row(
-                "pager_cache_hit_rate_pct",
-                format!("{:.2}", cache_hit_rate_pct),
-            ),
-        ];
-        Ok(ExecResult::Rows(rows))
-    }
-
-    fn emit_checkpoint_warning(&self, phase: &str, attempts: usize, error: &MuroError) {
-        let wal_path = self.wal.wal_path().display();
-        let wal_size = self
-            .wal
-            .file_size_bytes()
-            .ok()
-            .map(|v| v.to_string())
-            .unwrap_or_else(|| "unknown".to_string());
-        eprintln!(
-            "WARNING: checkpoint_failed phase={} attempts={} error=\"{}\" wal_path={} wal_size_bytes={}",
-            phase, attempts, error, wal_path, wal_size
-        );
-    }
-
-    fn try_checkpoint_truncate_once(&mut self) -> Result<()> {
-        #[cfg(test)]
-        if self.inject_checkpoint_failures_remaining > 0 {
-            self.inject_checkpoint_failures_remaining -= 1;
-            return Err(MuroError::Io(std::io::Error::other(
-                "injected checkpoint failure",
-            )));
-        }
-        self.wal.checkpoint_truncate()
-    }
-
-    fn try_checkpoint_truncate_with_retry(
-        &mut self,
-    ) -> std::result::Result<usize, (usize, MuroError)> {
-        let mut last_err = None;
-        for attempt in 1..=CHECKPOINT_MAX_ATTEMPTS {
-            if attempt > 1 {
-                std::thread::sleep(std::time::Duration::from_millis(1));
-            }
-            match self.try_checkpoint_truncate_once() {
-                Ok(()) => return Ok(attempt),
-                Err(e) => last_err = Some(e),
-            }
-        }
-        Err((
-            CHECKPOINT_MAX_ATTEMPTS,
-            last_err.unwrap_or_else(|| {
-                MuroError::Io(std::io::Error::other(
-                    "checkpoint truncate failed without error detail",
-                ))
-            }),
-        ))
-    }
-
-    #[cfg(test)]
-    fn inject_checkpoint_failure_once_for_test(&mut self) {
-        self.inject_checkpoint_failures_remaining = 1;
-    }
-
-    #[cfg(test)]
-    fn inject_checkpoint_failures_for_test(&mut self, count: usize) {
-        self.inject_checkpoint_failures_remaining = count;
-    }
-
-    #[cfg(test)]
-    fn set_checkpoint_policy_for_test(
-        &mut self,
-        tx_threshold: u64,
-        wal_threshold: u64,
-        interval_ms: u64,
-    ) {
-        self.checkpoint_policy = CheckpointPolicy {
-            tx_threshold,
-            wal_bytes_threshold: wal_threshold,
-            interval_ms,
-        };
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::crypto::aead::MasterKey;
-    use tempfile::TempDir;
-
-    fn test_key() -> MasterKey {
-        MasterKey::new([0x42u8; 32])
-    }
-
-    #[test]
-    fn test_commit_survives_checkpoint_failure_and_leaves_wal_for_recovery() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let wal_path = dir.path().join("test.wal");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&wal_path, &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        session.inject_checkpoint_failures_for_test(CHECKPOINT_MAX_ATTEMPTS);
-        let result = session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY, name VARCHAR)")
-            .unwrap();
-        assert!(matches!(result, ExecResult::Ok));
-
-        let wal_size = std::fs::metadata(&wal_path).unwrap().len();
-        assert!(
-            wal_size > 0,
-            "WAL should remain when checkpoint is injected to fail"
-        );
-
-        let mut db = crate::Database::open(&db_path, &test_key()).unwrap();
-        let rows = match db.execute("SELECT * FROM t").unwrap() {
-            ExecResult::Rows(rows) => rows,
-            _ => panic!("Expected rows"),
-        };
-        assert_eq!(rows.len(), 0);
-    }
-
-    #[test]
-    fn test_rollback_survives_checkpoint_failure_and_discards_uncommitted_data() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY, name VARCHAR)")
-            .unwrap();
-        session.execute("BEGIN").unwrap();
-        session
-            .execute("INSERT INTO t VALUES (1, 'alice')")
-            .unwrap();
-
-        session.inject_checkpoint_failures_for_test(CHECKPOINT_MAX_ATTEMPTS);
-        let result = session.execute("ROLLBACK").unwrap();
-        assert!(matches!(result, ExecResult::Ok));
-
-        let rows = match session.execute("SELECT * FROM t").unwrap() {
-            ExecResult::Rows(rows) => rows,
-            _ => panic!("Expected rows"),
-        };
-        assert_eq!(rows.len(), 0);
-    }
-
-    #[test]
-    fn test_commit_checkpoint_retries_transient_failure_and_clears_wal() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let wal_path = dir.path().join("test.wal");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&wal_path, &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        session.inject_checkpoint_failure_once_for_test();
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY, name VARCHAR)")
-            .unwrap();
-
-        let wal_size = std::fs::metadata(&wal_path).unwrap().len();
-        assert_eq!(
-            wal_size,
-            crate::wal::WAL_HEADER_SIZE as u64,
-            "transient checkpoint failure should be recovered by retry"
-        );
-    }
-
-    #[test]
-    fn test_retry_attempt_count_is_reported_on_transient_success() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        session.inject_checkpoint_failure_once_for_test();
-        let attempts = session.try_checkpoint_truncate_with_retry().unwrap();
-        assert_eq!(attempts, 2);
-    }
-
-    #[test]
-    fn test_auto_commit_survives_all_checkpoint_failures() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        // Inject failures that exhaust all retry attempts
-        session.inject_checkpoint_failures_for_test(CHECKPOINT_MAX_ATTEMPTS);
-
-        // Auto-commit should still succeed even though checkpoint fails
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY, name VARCHAR)")
-            .unwrap();
-
-        // Subsequent auto-commit should also succeed
-        session.inject_checkpoint_failures_for_test(CHECKPOINT_MAX_ATTEMPTS);
-        session
-            .execute("INSERT INTO t VALUES (1, 'alice')")
-            .unwrap();
-
-        // Data should be queryable
-        match session.execute("SELECT * FROM t").unwrap() {
-            ExecResult::Rows(rows) => assert_eq!(rows.len(), 1),
-            _ => panic!("Expected rows"),
-        }
-    }
-
-    #[test]
-    fn test_multiple_commits_each_retry_independently() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY, name VARCHAR)")
-            .unwrap();
-
-        // Each auto-commit gets its own independent retry cycle
-        for i in 0..5 {
-            session.inject_checkpoint_failure_once_for_test();
-            session
-                .execute(&format!("INSERT INTO t VALUES ({}, 'row_{}')", i, i))
-                .unwrap();
-        }
-
-        match session.execute("SELECT * FROM t").unwrap() {
-            ExecResult::Rows(rows) => assert_eq!(rows.len(), 5),
-            _ => panic!("Expected rows"),
-        }
-    }
-
-    #[test]
-    fn test_checkpoint_backoff_does_not_block_commit() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY)")
-            .unwrap();
-
-        // Even with backoff sleep, the commit completes successfully
-        session.inject_checkpoint_failure_once_for_test();
-        let start = std::time::Instant::now();
-        session.execute("INSERT INTO t VALUES (1)").unwrap();
-        let elapsed = start.elapsed();
-
-        // Backoff is 1ms, so the total should be well under 1 second
-        assert!(
-            elapsed < std::time::Duration::from_secs(1),
-            "Backoff should not significantly delay commit: {:?}",
-            elapsed
-        );
-
-        match session.execute("SELECT * FROM t").unwrap() {
-            ExecResult::Rows(rows) => assert_eq!(rows.len(), 1),
-            _ => panic!("Expected rows"),
-        }
-    }
-
-    #[test]
-    fn test_checkpoint_stats_tracked() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        // Initial stats should be zero
-        assert_eq!(session.checkpoint_stats().total_checkpoints, 0);
-        assert_eq!(session.checkpoint_stats().failed_checkpoints, 0);
-        assert_eq!(session.database_stats().deferred_checkpoints, 0);
-
-        // Successful checkpoint
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY)")
-            .unwrap();
-        assert_eq!(session.checkpoint_stats().total_checkpoints, 1);
-        assert_eq!(session.checkpoint_stats().failed_checkpoints, 0);
-
-        // Failed checkpoint (all retries exhausted)
-        session.inject_checkpoint_failures_for_test(CHECKPOINT_MAX_ATTEMPTS);
-        session.execute("INSERT INTO t VALUES (1)").unwrap();
-        assert_eq!(session.checkpoint_stats().total_checkpoints, 2);
-        assert_eq!(session.checkpoint_stats().failed_checkpoints, 1);
-        assert!(session.checkpoint_stats().last_failure_error.is_some());
-        assert!(session
-            .checkpoint_stats()
-            .last_failure_timestamp_ms
-            .is_some());
-        assert_eq!(session.database_stats().deferred_checkpoints, 0);
-    }
-
-    #[test]
-    fn test_checkpoint_can_be_deferred_by_tx_threshold() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let wal_path = dir.path().join("test.wal");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&wal_path, &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-        session.set_checkpoint_policy_for_test(3, 0, 0);
-
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY)")
-            .unwrap();
-        assert_eq!(session.database_stats().total_checkpoints, 0);
-        assert_eq!(session.database_stats().deferred_checkpoints, 1);
-        assert!(std::fs::metadata(&wal_path).unwrap().len() > crate::wal::WAL_HEADER_SIZE as u64);
-
-        session.execute("INSERT INTO t VALUES (1)").unwrap();
-        assert_eq!(session.database_stats().total_checkpoints, 0);
-        assert_eq!(session.database_stats().deferred_checkpoints, 2);
-
-        session.execute("INSERT INTO t VALUES (2)").unwrap();
-        assert_eq!(session.database_stats().total_checkpoints, 1);
-        assert_eq!(session.database_stats().deferred_checkpoints, 2);
-        assert_eq!(session.database_stats().failed_checkpoints, 0);
-        assert_eq!(
-            std::fs::metadata(&wal_path).unwrap().len(),
-            crate::wal::WAL_HEADER_SIZE as u64
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_triggered_by_wal_size_threshold_before_tx_threshold() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let wal_path = dir.path().join("test.wal");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&wal_path, &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY, name VARCHAR)")
-            .unwrap();
-
-        // Keep tx threshold high, but set WAL threshold low enough that large inserts trigger checkpoint.
-        session.set_checkpoint_policy_for_test(1000, 600, 0);
-        let checkpoints_before = session.database_stats().total_checkpoints;
-        session
-            .execute("INSERT INTO t VALUES (1, REPEAT('a', 2048))")
-            .unwrap();
-
-        assert_eq!(
-            session.database_stats().total_checkpoints,
-            checkpoints_before + 1
-        );
-        assert_eq!(session.database_stats().failed_checkpoints, 0);
-        assert_eq!(
-            std::fs::metadata(&wal_path).unwrap().len(),
-            crate::wal::WAL_HEADER_SIZE as u64
-        );
-    }
-
-    #[test]
-    fn test_show_checkpoint_stats_sql() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        // Execute a commit to get some stats
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY)")
-            .unwrap();
-
-        // Query stats via SQL
-        match session.execute("SHOW CHECKPOINT STATS").unwrap() {
-            ExecResult::Rows(rows) => {
-                assert_eq!(rows.len(), 4);
-                // First row: total_checkpoints
-                assert_eq!(
-                    rows[0].get("stat"),
-                    Some(&crate::types::Value::Varchar(
-                        "total_checkpoints".to_string()
-                    ))
-                );
-                assert_eq!(
-                    rows[0].get("value"),
-                    Some(&crate::types::Value::Varchar("1".to_string()))
-                );
-                // Second row: failed_checkpoints
-                assert_eq!(
-                    rows[1].get("value"),
-                    Some(&crate::types::Value::Varchar("0".to_string()))
-                );
-            }
-            _ => panic!("Expected rows from SHOW CHECKPOINT STATS"),
-        }
-    }
-
-    #[test]
-    fn test_show_database_stats_sql() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY)")
-            .unwrap();
-
-        match session.execute("SHOW DATABASE STATS").unwrap() {
-            ExecResult::Rows(rows) => {
-                assert_eq!(rows.len(), 18);
-                // Verify checkpoint stats
-                assert_eq!(
-                    rows[0].get("stat"),
-                    Some(&Value::Varchar("total_checkpoints".to_string()))
-                );
-                assert_eq!(rows[0].get("value"), Some(&Value::Varchar("1".to_string())));
-                // Verify commit_in_doubt_count present
-                assert_eq!(
-                    rows[4].get("stat"),
-                    Some(&Value::Varchar("commit_in_doubt_count".to_string()))
-                );
-                assert_eq!(rows[4].get("value"), Some(&Value::Varchar("0".to_string())));
-                // Verify freelist_sanitize_count present
-                assert_eq!(
-                    rows[7].get("stat"),
-                    Some(&Value::Varchar("freelist_sanitize_count".to_string()))
-                );
-                assert_eq!(rows[7].get("value"), Some(&Value::Varchar("0".to_string())));
-                assert_eq!(
-                    rows[10].get("stat"),
-                    Some(&Value::Varchar("deferred_checkpoints".to_string()))
-                );
-                assert_eq!(
-                    rows[11].get("stat"),
-                    Some(&Value::Varchar("checkpoint_pending_ops".to_string()))
-                );
-                assert_eq!(
-                    rows[12].get("stat"),
-                    Some(&Value::Varchar(
-                        "checkpoint_policy_tx_threshold".to_string()
-                    ))
-                );
-                assert_eq!(
-                    rows[13].get("stat"),
-                    Some(&Value::Varchar(
-                        "checkpoint_policy_wal_bytes_threshold".to_string()
-                    ))
-                );
-                assert_eq!(
-                    rows[14].get("stat"),
-                    Some(&Value::Varchar("checkpoint_policy_interval_ms".to_string()))
-                );
-                assert_eq!(
-                    rows[15].get("stat"),
-                    Some(&Value::Varchar("pager_cache_hits".to_string()))
-                );
-                assert_eq!(
-                    rows[16].get("stat"),
-                    Some(&Value::Varchar("pager_cache_misses".to_string()))
-                );
-                assert_eq!(
-                    rows[17].get("stat"),
-                    Some(&Value::Varchar("pager_cache_hit_rate_pct".to_string()))
-                );
-            }
-            _ => panic!("Expected rows from SHOW DATABASE STATS"),
-        }
-    }
-
-    #[test]
-    fn test_commit_in_doubt_increments_counter() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let wal_path = dir.path().join("test.wal");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&wal_path, &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        assert_eq!(session.database_stats().commit_in_doubt_count, 0);
-
-        // Set up a table first (needs a working pager)
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY)")
-            .unwrap();
-
-        // Inject write_page failure so next commit triggers CommitInDoubt
-        session
-            .pager_mut()
-            .set_inject_write_page_failure(Some(std::io::ErrorKind::Other));
-
-        let result = session.execute("INSERT INTO t VALUES (1)");
-        assert!(
-            matches!(&result, Err(MuroError::CommitInDoubt(_))),
-            "expected CommitInDoubt, got: {:?}",
-            result
-        );
-
-        // Counter must have been incremented
-        assert_eq!(session.database_stats().commit_in_doubt_count, 1);
-        assert!(session
-            .database_stats()
-            .last_commit_in_doubt_error
-            .is_some());
-        assert!(session
-            .database_stats()
-            .last_commit_in_doubt_timestamp_ms
-            .is_some());
-    }
-
-    #[test]
-    fn test_show_checkpoint_stats_still_works() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        // SHOW CHECKPOINT STATS should still return 4 rows (backward compat)
-        match session.execute("SHOW CHECKPOINT STATS").unwrap() {
-            ExecResult::Rows(rows) => {
-                assert_eq!(rows.len(), 4);
-            }
-            _ => panic!("Expected rows from SHOW CHECKPOINT STATS"),
-        }
-    }
-
-    #[test]
-    fn test_freelist_sanitize_stats_on_clean_open() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
-        let session = Session::new(pager, catalog, wal);
-
-        // Clean pager should have no sanitize stats
-        assert_eq!(session.database_stats().freelist_sanitize_count, 0);
-        assert_eq!(session.database_stats().freelist_out_of_range_total, 0);
-        assert_eq!(session.database_stats().freelist_duplicates_total, 0);
-    }
-
-    #[test]
-    fn test_stats_readable_on_poisoned_session() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let wal_path = dir.path().join("test.wal");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&wal_path, &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY)")
-            .unwrap();
-
-        // Poison the session via CommitInDoubt
-        session
-            .pager_mut()
-            .set_inject_write_page_failure(Some(std::io::ErrorKind::Other));
-        let result = session.execute("INSERT INTO t VALUES (1)");
-        assert!(matches!(&result, Err(MuroError::CommitInDoubt(_))));
-
-        // Regular queries must be rejected
-        let result = session.execute("SELECT * FROM t");
-        assert!(matches!(&result, Err(MuroError::SessionPoisoned(_))));
-
-        // SHOW DATABASE STATS must still work on poisoned session
-        match session.execute("SHOW DATABASE STATS").unwrap() {
-            ExecResult::Rows(rows) => {
-                assert_eq!(rows.len(), 18);
-                // commit_in_doubt_count should be 1
-                assert_eq!(
-                    rows[4].get("stat"),
-                    Some(&Value::Varchar("commit_in_doubt_count".to_string()))
-                );
-                assert_eq!(rows[4].get("value"), Some(&Value::Varchar("1".to_string())));
-                // Pager cache stats should be present
-                assert_eq!(
-                    rows[15].get("stat"),
-                    Some(&Value::Varchar("pager_cache_hits".to_string()))
-                );
-                assert_eq!(
-                    rows[16].get("stat"),
-                    Some(&Value::Varchar("pager_cache_misses".to_string()))
-                );
-                assert_eq!(
-                    rows[17].get("stat"),
-                    Some(&Value::Varchar("pager_cache_hit_rate_pct".to_string()))
-                );
-            }
-            _ => panic!("Expected rows from SHOW DATABASE STATS"),
-        }
-
-        // SHOW CHECKPOINT STATS must also work on poisoned session
-        match session.execute("SHOW CHECKPOINT STATS").unwrap() {
-            ExecResult::Rows(rows) => {
-                assert_eq!(rows.len(), 4);
-            }
-            _ => panic!("Expected rows from SHOW CHECKPOINT STATS"),
-        }
-    }
-
-    #[test]
-    fn test_read_only_query_select_does_not_create_wal_records() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let wal_path = dir.path().join("test.wal");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&wal_path, &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY, name VARCHAR)")
-            .unwrap();
-        session
-            .execute("INSERT INTO t VALUES (1, 'alice')")
-            .unwrap();
-
-        let wal_before = std::fs::metadata(&wal_path).unwrap().len();
-        let rows = session
-            .execute_read_only_query("SELECT id, name FROM t")
-            .unwrap();
-        let wal_after = std::fs::metadata(&wal_path).unwrap().len();
-
-        assert_eq!(rows.len(), 1);
-        assert_eq!(rows[0].get("id"), Some(&Value::Integer(1)));
-        assert_eq!(wal_before, wal_after, "read-only query must not append WAL");
-    }
-
-    #[test]
-    fn test_read_only_query_rejects_writes() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        let result = session.execute_read_only_query("INSERT INTO t VALUES (1)");
-        assert!(matches!(result, Err(MuroError::Execution(_))));
-    }
-
-    #[test]
-    fn test_read_only_query_in_explicit_tx_reads_uncommitted_state() {
-        let dir = TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-
-        let mut pager = Pager::create(&db_path, &test_key()).unwrap();
-        let catalog = SystemCatalog::create(&mut pager).unwrap();
-        pager.set_catalog_root(catalog.root_page_id());
-        pager.flush_meta().unwrap();
-        let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
-        let mut session = Session::new(pager, catalog, wal);
-
-        session
-            .execute("CREATE TABLE t (id BIGINT PRIMARY KEY, name VARCHAR)")
-            .unwrap();
-        session.execute("BEGIN").unwrap();
-        session
-            .execute("INSERT INTO t VALUES (1, 'alice')")
-            .unwrap();
-
-        let rows = session.execute_read_only_query("SELECT id FROM t").unwrap();
-        assert_eq!(rows.len(), 1);
-
-        session.execute("ROLLBACK").unwrap();
-        let rows = session.execute_read_only_query("SELECT id FROM t").unwrap();
-        assert!(rows.is_empty());
-    }
-}
